pyproject.toml
[project]
name = "norfab"

[tool.poetry]
name = "norfab"
version = "0.6.0"
description = "Network Automations Fabric [NorFab] - communication networks automations"
authors = ["Denis Mulyalin <d.mulyalin@gmail.com>"]
maintainers = ["Denis Mulyalin <d.mulyalin@gmail.com>"]
license = "Apache-2.0 license"
readme = "README.md"
homepage = "https://github.com/norfablabs/NORFAB"
repository = "https://github.com/norfablabs/NORFAB"
documentation = "https://docs.norfablabs.com"
keywords = ["Network", "Automation", "DevOps"]
classifiers = [
    "Topic :: Utilities",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
]

[tool.poetry.dependencies]
# Core Libs
python = ">=3.9,<3.12"
pyyaml = ">=6.0.1,<7.0.0"
pyzmq = ">=25.0.0,<27.0.0"
psutil = ">=6.0.0,<7.0.0"
tornado = ">=6.1.0,<7.0.0"
jinja2 = ">=3.1.3,<4.0.0"

# PICLE Client Dependencies
picle = { version = ">=0.8.7,<1.0.0", optional = true }
rich = { version = ">=13.0.0,<14.0.0", optional = true }
tabulate = { version = ">=0.9.0,<1.0.0", optional = true }
pydantic = { version = ">=2.3.0,<3.0.0", optional = true }
pyreadline3 = {version = ">= 3.0.0,<=4.0.0", optional = true }

# Nornir Service Dependencies
nornir = { version = "3.5.0", optional = true }
nornir-salt = { version = ">=0.22.2", optional = true }
cerberus = { version = "1.3.5", optional = true }
jmespath = { version = "1.0.1", optional = true }
napalm = { version = "5.0.0", optional = true }
ncclient = { version = "0.6.15", optional = true }
netmiko = { version = "4.5.0", optional = true }
nornir-napalm = { version = "0.5.0", optional = true }
nornir-netmiko = { version = "1.0.1", optional = true }
nornir-scrapli = { version = "2024.07.30", optional = true }
ntc-templates = { version = "4.1.0", optional = true }
paramiko = { version = "3.4.1", optional = true }
pygnmi = { version = "0.8.13", optional = true }
puresnmp = { version = "2.0.0", optional = true, extras = ['crypto'] }
requests = { version = ">=2.31.0,<=3.0.0", optional = true }
scrapli = { version = "2024.07.30", optional = true }
scrapli-community = { version = "2024.07.30", optional = true }
scrapli-netconf = { version = "2024.07.30", optional = true }
ttp = { version = "0.9.*", optional = true }
ttp-templates = { version = "0.3.*", optional = true }
xmltodict = { version = "0.13.0", optional = true }
lxml = { version = ">=4.9.3,<5.1.0", optional = true }
textfsm = { version = "1.1.3", optional = true }
N2G = { version = "0.3.*", optional = true }
dnspython = { version = "2.4.2", optional = true }
pythonping = { version = "1.1.4", optional = true }
robotframework = { version = "7.0", optional = true }
langchain = { version = "0.3.14", optional = true }
langchain-community = { version = "0.3.14", optional = true }
langchain-core = { version = "0.3.30", optional = true }
langchain-ollama = { version = "0.2.2", optional = true }
ollama = { version = "0.4.6", optional = true }

# Netbox Service Dependencies
diskcache = { version = "5.6.3", optional = true }
pynetbox = { version = "7.4.0", optional = true }

# FastAPI Service Dependencies
fastapi = { version = ">=0.115.0,<1.0.0", optional = true, extras = ["standard"] }
uvicorn = { version = "*", optional = true }
python-multipart = { version = "*", optional = true }

[tool.poetry.group.dev.dependencies]
black = "22.3.*"
pytest = { version = ">=7.1" }

# docs dependencies for extras definition
mkdocs = { version = "*" }
mkdocs-material = { version = "*" }
mkdocs-material-extensions = { version = "*" }
mkdocstrings = { version = "*", extras = ["python"] } 
pygments = { version = "*" }
pymdown-extensions = { version = "*" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project.entry-points."norfab.workers"]
"nornir" = "norfab.workers.nornir_worker:NornirWorker"
"netbox" = "norfab.workers.netbox_worker:NetboxWorker"
"fastapi" = "norfab.workers.fastapi_worker:FastAPIWorker"
"agent" = "norfab.workers.agent_worker:AgentWorker"
"workflow" = "norfab.workers.workflow_worker:WorkflowWorker"

[tool.poetry.scripts]
nfcli = 'norfab.utils:nfcli.nfcli'

[tool.poetry.extras]
nfcli = [
    "picle",
    "rich",
    "tabulate",
    "pydantic",
    "nornir-salt",
    "pyreadline3"
]
nornirservice = [
    "nornir",
    "nornir-salt",
    "cerberus",
    "jmespath",
    "napalm",
    "ncclient",
    "netmiko",
    "nornir-napalm",
    "nornir-netmiko",
    "nornir-scrapli",
    "ntc-templates",
    "paramiko",
    "pygnmi",
    "puresnmp",
    "requests",
    "scrapli",
    "scrapli-community",
    "scrapli-netconf",
    "ttp",
    "ttp-templates",
    "xmltodict",
    "lxml",
    "textfsm",
    "N2G",
    "dnspython",
    "pythonping",
    "tabulate",
    "pydantic",
]
netboxservice = [
    "requests",
    "pynetbox",
    "diskcache"
]
fastapiservice = [
    "fastapi",
    "uvicorn",
    "python-multipart"
]
robot = [
    "robot"
]
agentservice = [
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-ollama",
    "ollama",
]
full = [
    "picle",
    "rich",
    "tabulate",
    "pydantic",
    "nornir",
    "nornir-salt",
    "cerberus",
    "jmespath",
    "napalm",
    "ncclient",
    "netmiko",
    "nornir-napalm",
    "nornir-netmiko",
    "nornir-scrapli",
    "ntc-templates",
    "paramiko",
    "pygnmi",
    "puresnmp",
    "requests",
    "scrapli",
    "scrapli-community",
    "scrapli-netconf",
    "ttp",
    "ttp-templates",
    "xmltodict",
    "lxml",
    "textfsm",
    "N2G",
    "dnspython",
    "pythonping",
    "tabulate",
    "pydantic",
    "requests",
    "pynetbox",
    "robotframework",
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-ollama",
    "ollama",
    "fastapi",
    "uvicorn",
    "python-multipart",
]
